LayoutTests/imported/w3c/web-platform-tests/css/mediaqueries/mq-unknown-feature-custom-property-expected.txt


FAIL Serialization of <mf-name> : <mf-value> with custom property feature name and ident value assert_equals: expected "(--FOO: bar)" but got "not all"
